# packages/contracts/openapi/startup-graveyard.v1.yaml
openapi: 3.1.0
info:
  title: Startup Graveyard API
  version: 1.0.0
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: sg_access
      description: Host-only HttpOnly access cookie; public production uses the configured __Host- cookie name.
  schemas:
    CaseListItem:
      type: object
      required:
        - id
        - slug
        - companyName
        - industry
        - country
        - closedYear
        - summary
        - businessModelKey
        - foundedYear
        - totalFundingUsd
        - primaryFailureReasonKey
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        companyName: { type: string }
        industry: { type: string }
        country: { type: string, nullable: true }
        closedYear: { type: integer, nullable: true }
        summary: { type: string }
        businessModelKey: { type: string, nullable: true }
        foundedYear: { type: integer, nullable: true }
        totalFundingUsd: { type: number, nullable: true }
        primaryFailureReasonKey: { type: string, nullable: true }
...
    ListCasesResponse:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/CaseListItem' }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
...
    ReviewPublishReadiness:
      type: object
      required: [ready, evidenceCount, failureFactorCount, missing]
      properties:
        ready: { type: boolean }
        evidenceCount: { type: integer, minimum: 0 }
        failureFactorCount: { type: integer, minimum: 0 }
        missing:
          type: array
          items:
            type: string
            enum: [evidence_sources, failure_factors]
    ReviewListItem:
      type: object
      required:
        - id
        - caseId
        - slug
        - companyName
        - reviewStatus
        - assignedTo
        - decisionNote
        - createdAt
        - approvedAt
        - publishReadiness
      properties:
        id: { type: string, format: uuid }
        caseId: { type: string, format: uuid }
        slug: { type: string }
        companyName: { type: string }
        reviewStatus:
          type: string
          enum: [pending, changes_requested, approved, rejected]
        assignedTo: { type: string, nullable: true }
        decisionNote: { type: string, nullable: true }
        createdAt: { type: string }
        approvedAt: { type: string, nullable: true }
        publishReadiness:
          $ref: '#/components/schemas/ReviewPublishReadiness'
    ListReviewsResponse:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReviewListItem'
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0 }
...
    SourceSnapshotItem:
      type: object
      required:
        - id
        - sourceName
        - sourceUrl
        - finalUrl
        - httpStatus
        - contentType
        - title
        - excerpt
        - contentSha256
        - metadata
        - fetchedAt
        - createdAt
      properties:
        id: { type: string, format: uuid }
        sourceName: { type: string }
        sourceUrl: { type: string }
        finalUrl: { type: string }
        httpStatus: { type: integer, minimum: 100, maximum: 599 }
        contentType: { type: string, nullable: true }
        title: { type: string, nullable: true }
        excerpt: { type: string, nullable: true }
        contentSha256: { type: string }
        metadata:
          type: object
          additionalProperties: true
        fetchedAt: { type: string }
        createdAt: { type: string }
    ListSourceSnapshotsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SourceSnapshotItem'
...
    CopilotAdminMetrics:
      type: object
      required: [overview, feedbackEval, byPromptVersion, byFallbackReason, recentFlags, evals]
      properties:
        overview:
          $ref: '#/components/schemas/CopilotOverviewStats'
        feedbackEval:
          $ref: '#/components/schemas/CopilotFeedbackEval'
        byPromptVersion:
          type: array
          items:
            $ref: '#/components/schemas/CopilotPromptVersionStats'
        byFallbackReason:
          type: array
          items:
            $ref: '#/components/schemas/CopilotFallbackReasonCount'
        recentFlags:
          type: array
          items:
            $ref: '#/components/schemas/CopilotFlaggedRun'
        evals:
          $ref: '#/components/schemas/CopilotEvalAdminMetrics'
...
    CommercialAdminMetrics:
      type: object
      required: [subscriptions, billingFunnel, researchUsage, teamWorkspaces]
      properties:
        subscriptions:
          $ref: '#/components/schemas/SubscriptionAdminMetrics'
        billingFunnel:
          $ref: '#/components/schemas/BillingFunnelAdminMetrics'
        researchUsage:
          $ref: '#/components/schemas/ResearchUsageAdminMetrics'
        teamWorkspaces:
          $ref: '#/components/schemas/TeamWorkspaceAdminMetrics'
    PlatformRuntimeFeatureFlags:
      type: object
      required: [dbConfigured, adminEnabled, aiProvider, stripeEnabled, mockMode]
      properties:
        dbConfigured: { type: boolean }
        adminEnabled: { type: boolean }
        aiProvider:
          type: string
          enum: [anthropic, openai, none]
        stripeEnabled: { type: boolean }
        mockMode: { type: boolean }
    PlatformRecentFailedIngestionJob:
      type: object
      required: [id, sourceName, triggerType, errorMessage, createdAt, startedAt, finishedAt]
      properties:
        id: { type: string, format: uuid }
        sourceName: { type: string }
        triggerType: { type: string }
        errorMessage: { type: string, nullable: true }
        createdAt: { type: string }
        startedAt: { type: string, nullable: true }
        finishedAt: { type: string, nullable: true }
...
    PlatformWorkerMonitor:
      type: object
      required:
        [
          enabled,
          status,
          startDelayMs,
          pollIntervalMs,
          maxJobsPerTick,
          startedAt,
          lastTickStartedAt,
          lastTickCompletedAt,
          lastProcessedAt,
          lastProcessedJobId,
          lastProcessedSourceName,
          lastProcessedJobStatus,
          processedJobs,
          consecutiveErrors,
          lastError,
          lastStopAt,
          recentTicks,
        ]
      properties:
        enabled: { type: boolean }
        status:
          type: string
          enum: [disabled, idle, processing, error, stopped]
        startDelayMs: { type: integer, minimum: 0 }
        pollIntervalMs: { type: integer, minimum: 0 }
        maxJobsPerTick: { type: integer, minimum: 0 }
        startedAt: { type: string, nullable: true }
        lastTickStartedAt: { type: string, nullable: true }
        lastTickCompletedAt: { type: string, nullable: true }
        lastProcessedAt: { type: string, nullable: true }
        lastProcessedJobId: { type: string, format: uuid, nullable: true }
        lastProcessedSourceName: { type: string, nullable: true }
        lastProcessedJobStatus: { type: string, nullable: true }
        processedJobs: { type: integer, minimum: 0 }
        consecutiveErrors: { type: integer, minimum: 0 }
        lastError: { type: string, nullable: true }
        lastStopAt: { type: string, nullable: true }
        recentTicks:
          type: array
          items:
            type: object
            required:
              [
                startedAt,
                completedAt,
                outcome,
                processedCount,
                lastJobSourceName,
                lastJobStatus,
                error,
              ]
            properties:
              startedAt: { type: string }
              completedAt: { type: string }
              outcome:
                type: string
                enum: [processed, empty_queue, error]
              processedCount: { type: integer, minimum: 0 }
              lastJobSourceName: { type: string, nullable: true }
              lastJobStatus: { type: string, nullable: true }
              error: { type: string, nullable: true }
    PlatformSnapshot:
      type: object
      required:
        [
          createdAt,
          triggerType,
          mockMode,
          queuedCount,
          oldestQueuedAgeMinutes,
          runningCount,
          staleRunningCount,
          failedCount,
          completedLastHour,
          alertCount,
          criticalAlertCount,
          warningAlertCount,
          infoAlertCount,
          workerStatus,
          workerConsecutiveErrors,
          workerLastProcessedAt,
        ]
      properties:
        createdAt: { type: string }
        triggerType:
          type: string
          enum: [manual, scheduled]
        mockMode: { type: boolean }
        queuedCount: { type: integer, minimum: 0 }
        oldestQueuedAgeMinutes: { type: integer, minimum: 0, nullable: true }
        runningCount: { type: integer, minimum: 0 }
        staleRunningCount: { type: integer, minimum: 0 }
        failedCount: { type: integer, minimum: 0 }
        completedLastHour: { type: integer, minimum: 0 }
        alertCount: { type: integer, minimum: 0 }
        criticalAlertCount: { type: integer, minimum: 0 }
        warningAlertCount: { type: integer, minimum: 0 }
        infoAlertCount: { type: integer, minimum: 0 }
        workerStatus:
          type: string
          enum: [disabled, idle, processing, error, stopped]
        workerConsecutiveErrors: { type: integer, minimum: 0 }
        workerLastProcessedAt: { type: string, nullable: true }
    PlatformSnapshotTrend:
      type: object
      required:
        [
          sampleCount,
          oldestCapturedAt,
          latestCapturedAt,
          queuedCountDelta,
          oldestQueuedAgeDelta,
          alertCountDelta,
          failedCountDelta,
          workerConsecutiveErrorsDelta,
          maxQueuedCount,
          maxOldestQueuedAgeMinutes,
          maxAlertCount,
          maxFailedCount,
          maxWorkerConsecutiveErrors,
        ]
      properties:
        sampleCount: { type: integer, minimum: 0 }
        oldestCapturedAt: { type: string, nullable: true }
        latestCapturedAt: { type: string, nullable: true }
        queuedCountDelta: { type: integer, nullable: true }
        oldestQueuedAgeDelta: { type: integer, nullable: true }
        alertCountDelta: { type: integer, nullable: true }
        failedCountDelta: { type: integer, nullable: true }
        workerConsecutiveErrorsDelta: { type: integer, nullable: true }
        maxQueuedCount: { type: integer, minimum: 0 }
        maxOldestQueuedAgeMinutes: { type: integer, minimum: 0, nullable: true }
        maxAlertCount: { type: integer, minimum: 0 }
        maxFailedCount: { type: integer, minimum: 0 }
        maxWorkerConsecutiveErrors: { type: integer, minimum: 0 }
...
    PlatformSnapshotRollup:
      type: object
      required: [bucketSizeMinutes, bucketCount, buckets]
      properties:
        bucketSizeMinutes: { type: integer, minimum: 0 }
        bucketCount: { type: integer, minimum: 0 }
        buckets:
          type: array
          items:
            $ref: '#/components/schemas/PlatformSnapshotRollupBucket'
    PlatformSnapshotCadence:
      type: object
      required:
        [
          status,
          expectedIntervalMinutes,
          lastCapturedAt,
          lastScheduledCapturedAt,
          expectedNextSnapshotAt,
          minutesSinceLastSnapshot,
          minutesSinceLastScheduledSnapshot,
          overdue,
          missedIntervals,
        ]
      properties:
        status:
          type: string
          enum: [awaiting_baseline, healthy, overdue]
        expectedIntervalMinutes: { type: integer, minimum: 0 }
        lastCapturedAt: { type: string, nullable: true }
        lastScheduledCapturedAt: { type: string, nullable: true }
        expectedNextSnapshotAt: { type: string, nullable: true }
        minutesSinceLastSnapshot: { type: integer, minimum: 0, nullable: true }
        minutesSinceLastScheduledSnapshot: { type: integer, minimum: 0, nullable: true }
        overdue: { type: boolean }
        missedIntervals: { type: integer, minimum: 0 }
    PlatformSnapshotRegression:
      type: object
      required:
        [
          hasRegression,
          severity,
          regressionStreak,
          suppressed,
          suppressionReason,
          latestBucketStart,
          previousBucketStart,
          queuedDelta,
          oldestQueuedAgeDelta,
          alertCountDelta,
          failedCountDelta,
          workerConsecutiveErrorsDelta,
          reasons,
          recommendedActions,
        ]
      properties:
        hasRegression: { type: boolean }
        severity:
          type: string
          enum: [info, warning, critical]
        regressionStreak: { type: integer, minimum: 0 }
        suppressed: { type: boolean }
        suppressionReason: { type: string, nullable: true }
        latestBucketStart: { type: string, nullable: true }
        previousBucketStart: { type: string, nullable: true }
        queuedDelta: { type: integer, nullable: true }
        oldestQueuedAgeDelta: { type: integer, nullable: true }
        alertCountDelta: { type: integer, nullable: true }
        failedCountDelta: { type: integer, nullable: true }
        workerConsecutiveErrorsDelta: { type: integer, nullable: true }
        reasons:
          type: array
          items: { type: string }
        recommendedActions:
          type: array
          items: { type: string }
...
    PlatformSnapshotSuppressionSurface:
      type: object
      required:
        [
          windowHours,
          suppressedRegressionCount,
          unsuppressedRegressionCount,
          activeSuppressionReason,
          lastSuppressedBucketStart,
          topReasons,
        ]
      properties:
        windowHours: { type: integer, minimum: 0 }
        suppressedRegressionCount: { type: integer, minimum: 0 }
        unsuppressedRegressionCount: { type: integer, minimum: 0 }
        activeSuppressionReason: { type: string, nullable: true }
        lastSuppressedBucketStart: { type: string, nullable: true }
        topReasons:
          type: array
          items:
            $ref: '#/components/schemas/PlatformSnapshotSuppressionReasonBreakdown'
    PlatformSnapshotMetricsSurface:
      type: object
      required:
        [
          windowHours,
          coveredHours,
          snapshotCount,
          scheduledSnapshotCount,
          expectedScheduledSnapshotCount,
          cadenceAdherenceRate,
          regressionWindowCount,
          peakQueuedCount,
          peakOldestQueuedAgeMinutes,
          peakAlertCount,
          peakFailedCount,
          peakWorkerConsecutiveErrors,
        ]
      properties:
        windowHours: { type: integer, minimum: 0 }
        coveredHours: { type: integer, minimum: 0 }
        snapshotCount: { type: integer, minimum: 0 }
        scheduledSnapshotCount: { type: integer, minimum: 0 }
        expectedScheduledSnapshotCount: { type: integer, minimum: 0 }
        cadenceAdherenceRate:
          type: number
          minimum: 0
          maximum: 1
          nullable: true
        regressionWindowCount: { type: integer, minimum: 0 }
        peakQueuedCount: { type: integer, minimum: 0 }
        peakOldestQueuedAgeMinutes: { type: integer, minimum: 0, nullable: true }
        peakAlertCount: { type: integer, minimum: 0 }
        peakFailedCount: { type: integer, minimum: 0 }
        peakWorkerConsecutiveErrors: { type: integer, minimum: 0 }
...
    PlatformAdminMetrics:
      type: object
      required:
        [
          runtime,
          worker,
          recentSnapshots,
          snapshotCadence,
          snapshotTrend,
          snapshotRollup,
          snapshotRegression,
          snapshotSuppression,
          snapshotMetrics,
          ingestion,
          alertSummary,
          alerts,
        ]
      properties:
        runtime:
          type: object
          required: [service, env, nodeVersion, generatedAt, uptimeSeconds, features]
          properties:
            service: { type: string }
            env: { type: string }
            nodeVersion: { type: string }
            generatedAt: { type: string }
            uptimeSeconds: { type: integer, minimum: 0 }
            features:
              $ref: '#/components/schemas/PlatformRuntimeFeatureFlags'
        worker:
          $ref: '#/components/schemas/PlatformWorkerMonitor'
        recentSnapshots:
          type: array
          items:
            $ref: '#/components/schemas/PlatformSnapshot'
        snapshotCadence:
          $ref: '#/components/schemas/PlatformSnapshotCadence'
        snapshotTrend:
          $ref: '#/components/schemas/PlatformSnapshotTrend'
        snapshotRollup:
          $ref: '#/components/schemas/PlatformSnapshotRollup'
        snapshotRegression:
          $ref: '#/components/schemas/PlatformSnapshotRegression'
        snapshotSuppression:
          $ref: '#/components/schemas/PlatformSnapshotSuppressionSurface'
        snapshotMetrics:
          $ref: '#/components/schemas/PlatformSnapshotMetricsSurface'
        ingestion:
          type: object
          required:
            [
              queuedCount,
              oldestQueuedAgeMinutes,
              oldestQueuedSourceName,
              oldestQueuedTriggerType,
              completedLastHour,
              runningCount,
              staleRunningCount,
              staleThresholdMinutes,
              recentFailedCount,
              recentFailed,
              recentSucceeded,
              recentStale,
            ]
          properties:
            queuedCount: { type: integer, minimum: 0 }
            oldestQueuedAgeMinutes: { type: integer, minimum: 0, nullable: true }
            oldestQueuedSourceName: { type: string, nullable: true }
            oldestQueuedTriggerType: { type: string, nullable: true }
            completedLastHour: { type: integer, minimum: 0 }
            runningCount: { type: integer, minimum: 0 }
            staleRunningCount: { type: integer, minimum: 0 }
            staleThresholdMinutes: { type: integer, minimum: 0 }
            recentFailedCount: { type: integer, minimum: 0 }
            recentFailed:
              type: array
              items:
                $ref: '#/components/schemas/PlatformRecentFailedIngestionJob'
            recentSucceeded:
              type: array
              items:
                $ref: '#/components/schemas/PlatformRecentSucceededIngestionJob'
            recentStale:
              type: array
              items:
                $ref: '#/components/schemas/PlatformRecentStaleIngestionJob'
        alertSummary:
          type: object
          required: [critical, warning, info]
          properties:
            critical: { type: integer, minimum: 0 }
            warning: { type: integer, minimum: 0 }
            info: { type: integer, minimum: 0 }
        alerts:
          type: array
          items:
            $ref: '#/components/schemas/PlatformAlert'
    AdminStatsResponse:
      type: object
      required:
        - totalPublished
        - totalFundingUsd
        - totalDraft
        - avgFundingUsd
        - byIndustry
        - byCountry
        - byYear
        - byFailureReason
        - recentlyAdded
        - pendingReviews
        - ingestionStats
        - platform
        - commercial
        - copilot
      properties:
        totalPublished: { type: integer, minimum: 0 }
        totalFundingUsd: { type: number, minimum: 0 }
        totalDraft: { type: integer, minimum: 0 }
        avgFundingUsd: { type: number, minimum: 0 }
        byIndustry:
          type: array
          items:
            type: object
            required: [industry, count, totalFunding]
            properties:
              industry: { type: string }
              count: { type: integer, minimum: 0 }
              totalFunding: { type: number, minimum: 0 }
        byCountry:
          type: array
          items:
            type: object
            required: [country, count]
            properties:
              country: { type: string }
              count: { type: integer, minimum: 0 }
        byYear:
          type: array
          items:
            type: object
            required: [year, count]
            properties:
              year: { type: integer }
              count: { type: integer, minimum: 0 }
        byFailureReason:
          type: array
          items:
            type: object
            required: [reason, count]
            properties:
              reason: { type: string }
              count: { type: integer, minimum: 0 }
        recentlyAdded:
          type: array
          items:
            type: object
            required: [id, slug, companyName, createdAt]
            properties:
              id: { type: string, format: uuid }
              slug: { type: string }
              companyName: { type: string }
              createdAt: { type: string }
        pendingReviews: { type: integer, minimum: 0 }
        ingestionStats:
          type: object
          required: [pending, running, failed, completed]
          properties:
            pending: { type: integer, minimum: 0 }
            running: { type: integer, minimum: 0 }
            failed: { type: integer, minimum: 0 }
            completed: { type: integer, minimum: 0 }
        platform:
          $ref: '#/components/schemas/PlatformAdminMetrics'
        commercial:
          $ref: '#/components/schemas/CommercialAdminMetrics'
        copilot:
          $ref: '#/components/schemas/CopilotAdminMetrics'
    SchedulerJob:
      type: object
      required: [id, name, source_name, cron_expr, enabled]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        source_name: { type: string }
        cron_expr: { type: string }
        enabled: { type: boolean }
        last_run_at: { type: string, nullable: true }
        next_run_at: { type: string, nullable: true }
paths:
...
  /v1/cases:
    get:
      summary: List cases
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: industry
          schema: { type: string }
        - in: query
          name: country
          description: ISO 3166-1 alpha-2
          schema: { type: string, minLength: 2, maxLength: 2 }
        - in: query
          name: closedYear
          schema: { type: integer }
        - in: query
          name: businessModelKey
          description: 与 cases.business_model_key 精确匹配（小写）
          schema: { type: string, minLength: 1, maxLength: 100 }
        - in: query
          name: primaryFailureReasonKey
          description: 与 cases.primary_failure_reason_key 精确匹配（小写）
          schema: { type: string, minLength: 1, maxLength: 100 }
        - in: query
          name: sort
          description: >-
            有关键词 q 时默认按相关度：服务端若配置 OPENAI_API_KEY 且 case 有向量，
            则混合 pg_trgm 与向量余弦（各约 0.5 权重）；否则仅 trgm。可改为 updated_at。
            无 q 时默认按更新时间。
          schema:
            type: string
            enum: [relevance, updated_at]
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Case list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListCasesResponse'
...
  /v1/admin/reviews:
    get:
      summary: List reviews
      security:
        - BearerAuth: []
        - CookieAuth: []
        - AdminKey: []
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [pending, changes_requested, approved, rejected] }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Review list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListReviewsResponse'
        '401':
          description: Missing, expired, or invalid admin credential
        '503':
          description: Backing service unavailable
...
  /v1/admin/source-snapshots:
    get:
      summary: List recently captured source snapshots
      security:
        - BearerAuth: []
        - CookieAuth: []
        - AdminKey: []
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: sourceName
          schema: { type: string }
      responses:
        '200':
          description: Source snapshot list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSourceSnapshotsResponse'
        '401':
          description: Unauthorized
        '503':
          description: Backing service unavailable
  /v1/admin/cases:
    post:
      summary: Create draft case and pending review
      security:
        - BearerAuth: []
        - CookieAuth: []
        - AdminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug, companyName, summary, industryKey]
              properties:
                slug: { type: string, maxLength: 200 }
                companyName: { type: string, maxLength: 500 }
                summary: { type: string, maxLength: 20000 }
                industryKey: { type: string, maxLength: 100 }
                countryCode: { type: string, minLength: 2, maxLength: 2 }
                businessModelKey: { type: string, maxLength: 100 }
                foundedYear: { type: integer }
                closedYear: { type: integer }
                totalFundingUsd: { type: integer, minimum: 0 }
                primaryFailureReasonKey: { type: string, maxLength: 100 }
                assignedTo: { type: string, maxLength: 320 }
      responses:
        '200':
          description: Created draft + review
        '400':
          description: Invalid body
        '401':
          description: Unauthorized
        '409':
          description: Duplicate slug
        '503':
          description: Backing service unavailable
...
  /v1/admin/audit:
    get:
      summary: Recent admin audit events
      security:
        - BearerAuth: []
        - CookieAuth: []
        - AdminKey: []
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
      responses:
        '200':
          description: Audit list
        '401':
          description: Unauthorized
        '503':
          description: Backing service unavailable
  /v1/admin/stats:
    get:
      summary: Admin dashboard aggregate statistics
      security:
        - BearerAuth: []
        - CookieAuth: []
        - AdminKey: []
      responses:
        '200':
          description: Admin metrics payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStatsResponse'
        '401':
          description: Unauthorized
        '503':
          description: Backing service unavailable
...
  /v1/admin/ingestion-jobs:
    get:
      summary: Recent ingestion jobs
      security:
        - BearerAuth: []
        - CookieAuth: []
        - AdminKey: []
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
        - in: query
          name: status
          schema:
            type: string
            enum: [queued, running, succeeded, failed]
      responses:
        '200':
          description: Job list
        '401':
          description: Unauthorized
        '503':
          description: Backing service unavailable
    post:
      summary: Enqueue ingestion job (queued status)
      security:
        - BearerAuth: []
        - CookieAuth: []
        - AdminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourceName, triggerType]
              properties:
                sourceName: { type: string, maxLength: 200 }
                triggerType: { type: string, maxLength: 80 }
                payload: { type: object, additionalProperties: true }
      responses:
        '200':
          description: Created job id
        '400':
          description: Invalid body
        '401':
          description: Unauthorized
        '503':
          description: Backing service unavailable
...
  /v1/admin/scheduler:
    get:
      summary: List scheduled ingestion/maintenance jobs
      security:
        - BearerAuth: []
        - CookieAuth: []
        - AdminKey: []
      responses:
        '200':
          description: Scheduler jobs
          content:
            application/json:
              schema:
                type: object
                required: [jobs]
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/SchedulerJob'
        '401':
          description: Unauthorized
        '503':
          description: Backing service unavailable
...